rest_api/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
SQLAlchemy==1.4.46
python-decouple==3.7
Flask-Cors==3.0.10
psycopg2==2.9.5
Flask-Migrate